+++ src/add-folder.tsx
@@ -1,5 +1,0 @@
-import AddFolderForm from "./components/AddFolderForm";
-
-export default function Command() {
-  return <AddFolderForm />;
-}